requirements.txt
streamlit>=1.28.0
easyocr>=1.7.0
google-genai>=0.2.0
opencv-python-headless>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
